Revert some changes:
- Make targets explicit b7ad08df84f85429c978ec855d40b2fd4c3e213b.
- Get rid of Vfile      e73fd90b65c912045a22245ece86978f2d88b15a.

Without vfile we need some new concepts to avoid parsing the requires
file multiple times and with vfile it's annoying to specify the
dependencies by hand.

Will leave that for future work. Just use memoize where it make sense,
for instance when we read the result from only the current directory
(for instance the ocamldep stuff).