WIP: strict inclusion of library
Sometimes, in order to enforce abstractions, we might want to build a
library without allowing access to all the `.cmi` of the 'recursive'
libraries it depend upon. For instance, we might want to call the
OCaml compiler with only the `-I` of the explicitly listed libraries
and not the `-I` of their recursives closures.

This patch allow this behaviour by introducing an optional
`!` as first character of a dependent library.

The default behaviour is not changed and obviously this patch
preserves the recursive closure of dependent libraries when calling
the executable linkers.