Handle CRLF files correctly in let-syntax
Previously a CRLF source file would be re-emitted with \r\n translated
to \r\r\n which affected the fragments in src/preprocessing.ml. The
lexer is now altered so that \r is never written to the output channel
so on Windows the result will always be CRLF encoded and on Unix the
result will always be LF encoded, regardless of the input.

Signed-off-by: David Allsopp <david.allsopp@metastack.com>