configurator: add an unsigned int type for C symbol imports
The current `C_define.Type.Int` supports signed integers since
the fix in #1344. This fix meant that some C expressions cannot
be evaluated such as `sizeof(expr)` since those are not valid
in cpp.

This changeset adds a new `Uint` type switch that is expressly
for the purpose of evaluating unsigned C integers only. This
allows for constructs such as `sizeof(expr)`.

This changeset lets the Ctypes library completely use Configurator
for all of its compile-time tests (ocamllabs/ocaml-ctypes#574)

fixes #1720

Signed-off-by: Anil Madhavapeddy <anil@recoil.org>