Track original source location in lib_info/dune_package
This allows to communicate original source locations so we can construct
`.merlin` files with `S` directives pointing right to them for installed
libs.

This is useful when lib is opam-pinned or esy-linked. I think we can set
`orig_src_dir` to `None` for all other installed packages (based on
profile?).

Signed-off-by: Andrey Popp <8mayday@gmail.com>