[configurator] Return pkg-config error to clients of new query API.
It is very convenient for configuration tools to be able to provide a
better error message to the user when pkg-config fails.

This was suggested by @garrigue.

We take advantage of the newly [undocumented] API in 1.7.2 to modify
its return type, but this could be a problem. If so, we will have to
introduce yet another function if we want this functionality.

We also update `src/configurator/dune` to include transitive
dependencies as suggested by @rgrinberg

Signed-off-by: Emilio Jesus Gallego Arias <e+git@x80.org>