[coq] Add support for native-mode compilation.
We add a `(mode ...)` field to the `(coq.theory ...)` stanza that when
set to `native` will have the Coq compiler to generate and install
"native" objects, that is to say, `.cmxs` libraries containing an
ML-level version of the corresponding Coq module.

As of today, Coq itself does call the `ocamlfind ocaml` to compile the
object files, this restriction may be lifted in the future however it
is not easy as tactics do have access to "Just-In-Time" native
compilation too.

The patch itself should have been straightforward except for a big
problem due to the way Coq outputs native targets and the limitations
of Dune.

By default, `coqc -native-compute on` will generate the OCaml-level
objects files under `.coq-native/`, and the regular `.vo` object in
the build dir. This scheme doesn't work well with the restriction than
a rule can only produce targets in the same directory.

To workaround this limitation, we pass a special flag to Coq so
`.coq-native` is dropped from the output path; at install time we do
set the rules up so indeed the objects will be installed under
`.coq-native` for compatibility. This means this feature will only
work in Coq >= 8.12.1

Moreover, the non-standard location of native objects defeats Coq's
default native include path, so we must pass the additional
non-standard include paths to `coqc`.

Note that native compilation is disabled by default in the `dev`
profile, as to speed up build. Please use the release or a custom
profile if you want to avoid this.

Signed-off-by: Emilio Jesus Gallego Arias <e+git@x80.org>