Path Dir_status to use Path.Build.t where possible
Dir_status is only kept for build directories and we use the correct
path type to reflect that.

This PR also addresses a faulty like looking check for the project root. The
project root is a source path, and the dir whose status we're getting is a build
path. There was no way those would be equal, the build dir's context has to be
stripped first.

Signed-off-by: Rudi Grinberg <rudi.grinberg@gmail.com>