Remove unused abbreviation "jsoo"
There is currently no value in Introducing the "jsoo" abbreviation.

The term "jsoo" is used exactly once (excluding the sentence where the abbreviation is introduced), while "js_of_ocaml" is used 11 times. Moreover, at most places in the doc, "js_of_ocaml" can't be easily replaced with "jsoo", as it refers to either the command line tool, or to the dune configuration, where it is named js_of_ocaml and not jsoo.

So it makes more sense to just remove "jsoo", which shortens the documentation by a whole sentence and expands the one occurrence.

Signed-off-by: Volker Diels-Grabsch <v@njh.eu>