Implement compound dependency rules for vlibs
Instead of taking vlib and impl dep graphs separately and then merging
them, we now have a higher level dependency procedure that is aware of
the special treatment implementations of virtual libraries need.

This opportunity is also take to move the special logic for alias and
wrapped_compat modules into one place.

Signed-off-by: Rudi Grinberg <rudi.grinberg@gmail.com>