Silence merlin approximation warning in lang < 1.9
Closes #2399

The merlin approximation warning says that it can be silenced by adding
(allow_approximate_merlin) in dune-project, but it is not an option if
the lang version is < 1.9.

To make sure that this warning is not actionable, it is emitted only if
the lang version in the corresponding project is >= 1.9.

Signed-off-by: Etienne Millon <me@emillon.org>