Mention Coq in the overview.
Since #2531, there is no longer a Coq section in the documentation,
and since #2535 there is no longer a Coq sub-section in the "Other
topics" section.  Coq's documentation is only available through its
dedicated stanzas.  This means that it is much harder to discover that
Dune supports not just OCaml and Reason, but also Coq.  This is an
attempt to fix that, although it might be controversial, especially
given that Coq's support is still very much a work-in-progress.

Co-authored-by: Nicolás Ojeda Bär <n.oje.bar@gmail.com>

Signed-off-by: Théo Zimmermann <theo.zimmermann@univ-paris-diderot.fr>