Adds unit test reproducing problem with rule rerun.
Adds unit test showing that a rule depending on a glob is not rerun
when the glob changes but the set of files it matches to does not.

This is a problem because building a target with invalid rule
can prevent the rule from being rerun even after the rule is fixed.

Signed-off-by: Jakub Staron <jstaron@janestreet.com>