Remove the `Universe` "fact" from the dependency trace.
The `Universe` fact can't be used to force rebuild because it doesn't
actually carry the digest of the universe, so we already have to rely
on a different mechanism to force rebuild (build_system.ml calls
`Dep.Set.has_universe` explicitly).

Since it serves no purpose and makes no sense, let's remove it.

Signed-off-by: Arseniy Alekseyev <aalekseyev@janestreet.com>