Move all cycle checks to [Dep_stack]
Previously, we'd rely on the [St_initializing] constructor to detect
cycles. In fact, this isn't necessary and all cycle checks can be done
in the dep stack. All that is required is for the unique library Id to
be available. Since the id is now only a function of the name + path,
this is now possible.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>