Remove [Build.Lazy_no_targets] (#3062)
We used to need a way to hide certain [Build.t] subexpressions
behind [Lazy_no_targets] constructors. These constructors were
not explored when computing the list of build targets (with
[Build.targets]) and allowed us to avoid a cyclic dependency
between computing the list of targets and resolving [If_file_exists]
queries (hidden behind [Lazy_no_targets]).

Thanks to #3052, we no longer need to traverse [Build.t] expressions
to compute the list of build targets, and we can therefore remove
this bit of complexity.

Note that laziness here is used only to avoid cycles, i.e. it is not
a performance optimisation, because we do evaluate everything when
computing the list of static dependencies with [Build.static_deps].

I checked that there is no degradation of performance or memory usage.

Signed-off-by: Andrey Mokhov <amokhov@janestreet.com>