[#3007] Add common commands into the synopsis
This is a very minimal approach to resolve #3007. It relies on the
command names to be self-documenting, supported by the fact that more
extensive information is available just below.

It aims to be consistent with the purposes of the SYNPOSIS section
according to MAN(1) while stile providing useful hints to newer users.

Signed-off-by: Shon Feder <shon.feder@gmail.com>