Remove ad-hoc memoization in [Env_node] (#3168)
This removes mutation from [Env_node], switching to [Memo.Lazy]. 

The logic is also slightly simplified by factoring out some common
patterns (e.g. [inherited]). The resulting code is still rather
convoluted due to the proliferation of lazy computations but
hopefully the new state is still an improvement.

When implementing, we noticed that tests were not very good (the
first draft was wrong yet passed the testsuite), so we've also added
a test to make sure we use the right expanders.

Signed-off-by: Andrey Mokhov <amokhov@janestreet.com>

Co-authored-by: Arseniy Alekseyev <aalekseyev@janestreet.com>