[coq] Fix install path for theories with a composed name.
Theories with names such as `mathcomp.ssreflect` should be installed
in `mathcomp/ssreflect` for the common pattern `From mathcomp Require
Import ...` to work.

Thanks to Shachar Itzhaky for testing and reporting this issue.

The use of the recursive include is problematic upstream and a revamp
of its semantics is planned.

Signed-off-by: Emilio Jesus Gallego Arias <e+git@x80.org>