Split cram sanitation from public library
In the current setup, dune's output sanitation logic leaks into cram.
This is not desirable because we are planning to make cram a public
library.

This PR attempts to solve this problem by introducing a tiny cram
library where the user can craft his own cram binary and set the
sanitizer however he wishes.

The advantage is that the sanitizer can be composed with sanitizer
already defined by cram, and that everything is done in one process. The
cost is a tiny library introduced.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>