push basic api refactor (NON-COMPILE)
The edits for this commit were constructed programmatically, by first
generating a dependency graph for the contents of what used to be
src/dune, then moving all transitive dependencies of the module
[Build_system] into the folder [src/engine], leaving the remainder as
[src/build_rules].

The changes in this commit can be summed up as follows:

- If a changed file is in [src/engine], it is unchanged beyond the
folder rename.
- If a changed file is in [src/build_rules], it has been prepended with
the line [open! Engine], but is otherwise unchanged.
- Otherwise, files have been changed to refer to [Engine] rather than
[Dune].

Ideally, the contents of the [src/engine] folder would be the "public
API" as available to the rules, dispatching to the nitty-gritty of
actually building the targets. However, the current setup is a bit too
entangled for that to be done easily (mostly due to various "back-end"
modules needing access to "front-end" modules for error-reporting
purposes, such as [Findlib] requiring [Context]).

This commit does not build on my box, as it does not properly account
for @rgrinberg's Cramtest additions (as the splitting script was working
off an old dependency graph).

Signed-off-by: Cameron Wong <cwong@janestreet.com>