cram directory tests: skip empty directories
In the case of cram directory tests (directories with a name that ends
in .t), there's a check that they contain a `run.t` file. When that is
not the case, the build errors out.

It can cause problems because empty directories are not tracked by git.
An update can leave empty directories behind, which can make builds fail
in a mysterious way.

This relaxes the existing check by skipping empty directories.

Signed-off-by: Etienne Millon <me@emillon.org>