Use Lib_info to generate install archives
Lib_info is what we us to read installed metadata, so this is what we
should use to generate the list of installed files.

As a bonus, most of the information that's available in Lib_info is
already evaluated, so we can drop some guards in the install_rules and
move them all the creation of Lib_info.t.

Consequently, this fixes the generation of META/dune-package files.
Previously, archives that were disabled were still included in these
files. Now they are correctly excluded.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>