Fix project lookup for libraries
Whenever we know the library is defined in the context, we can look up
its project directly. There's no need to reverse engineer it from the
packages, as this would be incorrect for vendored packages in the same
workspace. For example:

```
vendor/
 foo/
  foo.opam
foo.opam
```

The current code would wrongly determine that libraries in either foo
packages are different in the same project.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>