Produce & Promote man page with sphinx
This allows us to ship the entire dune manual as one big man page. That
already makes it pretty useful to browse docs offline.

It's possible to split the man pages up, but this is just an experiment
for now.

I'm not a huge fan of checking in the man pages like this. Perhaps
there's a way to do this step as part of dune-release?

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>