Re-do build cancellation
Previously, build cancellation relied on file change events interrupting
`Fiber.run`. This will not work if we want rpc to stay uninterrupted.

We change cancellation to work by setting the status of the scheduler to
be `Cancelling_build`. Once the status is set, we kill running build
processes and forbid launching new ones to await the termination of the
build. Then, we re-run the build.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>