Never pass -nodynlink option
This is an optimization when PIC executables are not used, but this
optimization is disabled becauses it causes errors on arm32. Most
distributions are going in the direction of requiring PIC, and
`-nodynlink` might go away (see ocaml/ocaml#8867), so the supporting
code in dune (which is bypassed in most cases) can be removed.

Closes #4069
Closes #2527

Signed-off-by: Etienne Millon <me@emillon.org>