clarify the location of the documentation
Before this PR `make doc` ends with:

> The HTML pages are in _build.

But if you go looking for the documentation in `_build` you are
doomed, because it is not there, it is actually in
`doc/_build` (with no message whatsoever suggesting that the line
below was printed from the `doc/` subdirectory).

Now the message says:

> The HTML pages are in doc/_build.

which gives you a chance of locating the documentation.

Signed-off-by: Gabriel Scherer <gabriel.scherer@gmail.com>