Assert the non-existence of zombies (#4255)
It used to be possible for a computation to finish with a cycle error and remain
in the [Considering] state forever, becoming a "zombie" computation. To
distinguish between "current" and "zombie" computations, we stored the [run] in
which the computation had started. In this way, before subscribing to an
[Ivar.t] from the [completion], we could check if it corresponded to the current
run, and if not, start a new computation.

With better error-handling introduced in #4015, we believe such zombie
computations are no longer possible. In this PR, we make the function
[currently_considering] throw a [Code_error] if it encounters a zombie.

Once we have convinced ourselves that there are no zombies out there, we can
remove the [run] from the [Considering] state.

Signed-off-by: Andrey Mokhov <amokhov@janestreet.com>