Use Unix.create_process instead of fork
In 4.12, create_process now uses posix_spawn instead of fork. Resulting
in some performance benefits. In my unscientific benchmarking, running
/bin/true in a loop of 1000 iterations is 1.37 vs. 1.82 seconds in favor
of create_process on a mac. The difference is likely smaller on Unix,
since forking on mac is quite slow. In any case, we already avoid fork
on Windows. So why not keep things both consistent and performant and
switch to create_process everywhere.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>