Add a new implementation of the shared cache and start using it (#4443)
This PR switches Dune to use well-tested and optimised Jenga's implementation of the
shared cache. Apart from using the new logic with fewer races (which now matches the
documentation in [doc/dev/cache.md]), there are a few other changes:

* We no longer plumb store/restore requests through the daemon, i.e. we switch to using
the direct mode exclusively. This simplifies Dune quite a bit. The only cost is that in the
copying mode, we currently need to wait until the artifacts are copied to the cache before
proceeding with the build.

* The cache metadata format has been simplified: we no longer store full paths to build
artifacts, it is sufficient to store the names of files. Hence the bump of the version to [meta/v5].

Some changes are left for follow-up PRs to keep this one as small as possible:

* Remove the cache daemon.
* Refactor the cache trimmer to use the new cache storage libraries.
* Change the cache configuration format.
* Reuse Jenga's shared cache test suite.
* Implement the cloud cache functionality.
* Update developer and user docs.

Signed-off-by: Andrey Mokhov <amokhov@janestreet.com>