Add new configuration (skip_ppx_kind <bool>)
Recent version of dune language check that ppx rewriters
have a psecific kind. In case when a rewriter is being
dynamically extended by new plugins, it leads to an error:

```
Error: Ppx dependency on a non-ppx library "GT.syntax.show". If
"GT.syntax.show" is in fact a ppx rewriter library, it should have (kind
ppx_rewriter) in its dune file.
```

This PR addes a new switch with following sematics

* `(skip_ppx_kind true)` skips checking rewriters' kinds
* `(skip_ppx_kind false)` (default) do the check
* without explicit specification it defaults to `false`

Signed-off-by: Dmitrii Kosarev <Dmitrii.Kosarev@pm.me>