Use arrays for storing Memo node dependencies (#4687)
[Cached_value.deps] stores a list of dependencies [deps]. We
construct this list when creating a cached value but traverse
it on every rebuild (i.e. more often), so it makes sense to
make this traversal faster by switching to an array and thus
avoiding pointer chasing. This also saves some memory (roughly
two words per node dependency).

My benchmarks show a noticeable improvement (of the order of 10%),
given that this is a hot path for incremental rebuilds. 

Signed-off-by: Andrey Mokhov <amokhov@janestreet.com>