[install_rules] Allow for stanzas to install files in different packages.
In some cases, it is convenient for stanzas to produce objects that
will be installed under different packages.

For example, in Coq we may want to install the costly native files in
a `foo-native` package, or distributions such as Debian like to split
packages in `-doc`, `-dev`, and `-dbg` variants.

This PR modifies the code so this is possible.

This code is cherry-picked from #4750 (which would be the first user
of the feature), but I am submitting separately as I think the change
deserves more discussion.

For example, this implies that `Dune_file.stanza_package stanza` is not
canonical anymore. Likely, a better modelling would be to introduce
the notion of "sub-package".

The code changes themselves don't make me super-happy, consider them a
draft. If you folks are OK with the idea, I can improve it so the
common case of one stanza installing files only for one package is
better optimized.

Signed-off-by: Emilio Jesus Gallego Arias <e+git@x80.org>