[coq] Allow to install Coq native files under a different package.
The code changes doesn't make me super-happy, tho I think that in
general it makes sense for stanzas to produce object files that do
belong to multiple packages, as for example it is common in some
upstream packagers to allow to install debug and doc files into a
different package.

However this implies that `Dune_file.stanza_package stanza` is not
canonical anymore.

Likely to model this better, we may need to define the notion of
sub-packages of a stanza / package, but awaiting for feedback.