Stop reusing DAG nodes for restore and compute (#4797)
Change Memo to stops reusing DAG nodes for different phases.

This change introduces a performance regression, in particular, we would previously create
only one stack frame for nodes with cutoff during the incremental zero rebuilds but now we
introduce two such nodes: we first call [compute] (creating a stack frame), which then calls
[restore_from_cache] (creating another stack frame). 

We think this patch still moves us in the right direction, so we are accepting the regression
for now. We have some ideas about optimising the call stack representation that should
hopefully bring the performance back. These changes are better done in a separate PR.

Signed-off-by: Andrey Mokhov <amokhov@janestreet.com>

Co-authored-by: Arseniy Alekseyev <aalekseyev@janestreet.com>