Push the [has_cutoff] branch into the compute phase (#4836)
We currently branch on [dep.has_cutoff] for every [dep], even if that [dep] 
can be successfully restored from the cache and doesn't need to be recomputed.
Such branches are costly because they are hard to predict, so it's best to make
them only when necessary. This PR pushes the branch into the computation phase.

Microbenchmarks look good but not spectacular. Anyway, the restore from cache scenarios are improved, which is what we mostly care about for speeding up incremental builds.

Signed-off-by: Andrey Mokhov <amokhov@janestreet.com>