Fix progress display
The progress display wasn't displayed during the first build. It was
only displayed from the second build in polling mode. This was because
of a `Status_line.set_constant None` in the function that scanned the
source tree.

The fix is to not clear the status line in this function, and instead
use a "status line overlay".

Signed-off-by: Jeremie Dimino <jeremie@dimino.org>