[tests] Update plugin tests for Coq 8.14
Coq 8.14 split the ML libraries to a `coq-core` package, actually with
the goal to be able to make Coq's install much lighter for users like Dune.

Unfortunately we didn't add a compat layer as anyways the ML API is
not 100% stable. Sorry to all for the inconvenience.

By the way, this makes me wonder if we should use an OPAM lock file
for Dune development?

Signed-off-by: Emilio Jesus Gallego Arias <e+git@x80.org>