Warn when lang versions have an ignored suffix
Syntax versions (the string after `lang dune` for example) are parsed
with `Scanf.sscanf`. This only parses the longest matching prefix and
ignored the rest of the string. So `lang dune 2.3.0` is silently
interpreted as `2.3` and so is `lang dune 2.4suffix`.

This PR adds a warning when there is a suffix. An error could break
existing projects.

The implementation adds a `%s` to match the suffix. This specifier stops
on whitespace characters, but these can not appear in valid atoms.

Signed-off-by: Etienne Millon <me@emillon.org>