Warn when lang versions have an ignored suffix (#5040)
Syntax versions (the string after `lang dune` for example) are parsed
with `Scanf.sscanf`. This only parses the longest matching prefix and
ignored the rest of the string. So `lang dune 2.3.0` is silently
interpreted as `2.3` and so is `lang dune 2.4suffix`.

This PR adds a check for suffixes. It is a warning with lang < 3.0 and an
error with lang >= 3.0, in order to not break existing projects.

The implementation adds a `%s` to match the suffix. This specifier stops
on whitespace characters, but these can not appear in valid atoms.

Signed-off-by: Etienne Millon <me@emillon.org>