Test Dune RPC behaviour in the case of re-triggering (#5103)
I was super confused by seeing [Failure] from the RPC when the build was in
fact merely re-triggering. It costed me quite a bit of time to figure out
what's going on. So, I'm adding a test to demonstrate the current behaviour
along with a CR-someday to fix it if everyone agrees.

Signed-off-by: Andrey Mokhov <amokhov@janestreet.com>