Stop caching unused fields in [Fs_cache.Reduced_stats] (#5203)
The [Fs_cache.Reduced_stats] record is too large and some of its fields are
currently unused. They cause unnecessary restarts for Dune, so I am removing
them. The [mtime] field is particularly problematic because it changes for a
directory when we create temporary files inside it, which I believe is the
cause for #5170.

I originally included these fields into [Fs_cache.Reduced_stats] because
they appeared important for the [Cached_digest] module. However,
[Cached_digest] obtains them via untracked file-system access and we are
not planning to change that for now. This means, there is no reason to
keep them in [Fs_cache.Reduced_stats].

This fixes a problem in an existing file-watching test and also improves
builds in my manual testing.

Signed-off-by: Andrey Mokhov <amokhov@janestreet.com>