Fix handling of "too many links" error in the shared cache (#5284)
We've been occasionally getting some shared cache errors mentioning [EEXIST] and [link] when running large
builds internally. I've finally managed to debug this: apparently the code path that handles the "too many links"
error in the shared cache had a silly error which rendered this code path useless. This PR fixes the error and
adds some comments.

You might wonder: how can 65000 links be not enough for the shared cache? It turns out we build a lot of
empty files which all link to the same entry in the shared cache. These are mostly empty lists of dependencies
stored in [.d] files.

I tested this manually. I can't come up with a reliable and not super slow test to add to CI.

Signed-off-by: Andrey Mokhov <amokhov@janestreet.com>