Minor cleanup of code for ``dune describe``
Made explicit the type definitions that represent the information that
is dumped about the workspace. Translation to the Dyn.t type happens
as a second stage.

The goal of this commit is hopefully to make users' life easier in
decoding the S-expressions into meaningful contents, and to make it
easier to extend the features of ``dune describe`` in the future.

Signed-off-by: Benoît Montagu <benoit.montagu@inria.fr>