[coq top] Test case for problems with path / dir computation in dune coq top
After some more testing, I have found problems finding the directory /
file for `dune coq top`:

```ocaml
        let coq_file_build =
          let p = Common.prefix_target common coq_file_arg in
          Path.Build.relative context.build_dir p
        in
        let dir =
          let dir = Filename.dirname coq_file_arg in
          let p = Common.prefix_target common dir in
          Path.Build.relative context.build_dir p
        in
```

`dune coq top` will also fail on project that have a `theories/a/b.v`
layout, such as in Coq for example, when doing:

```
dune coq top theories/Reals/Real.v
```

in the above snippet, `dir` is set to `theories/Reals` which is
incorrect, as it should be the root of where the stanza is placed.

We need to think more about this, and fix it because it is currently
only usable on flat layouts.

Let's discuss in this PR.