Attach bin/main.exe to the dune package
When dune is being built by another dune, it will refer to the
package this executable is attached to to fetch the right build-info.

This mechanism is not used when bootstrapping.

A way to see this change is to set a version number in `dune-project`
and then to use `dune build` / `dune install` to install dune itself.

Closes: #5554
Signed-off-by: Etienne Millon <me@emillon.org>