[coq_rules] Refactor boot flag handling.
Instead of doing an special case in Coq_lib, we handle the boot
library directly in the flags.

This allows us to handle coqdoc-specific flags related to `-boot`, and
is more in line with what we expect the future theory model will be.

Downside is that `-boot` is now not included in the closure dep, but
that seems fine, actually we should forbid `(boot)` libraries from
declaring `(theories ...)` deps, but as we are getting rid of it
eventually maybe it is not worth the effort.

Signed-off-by: Emilio Jesus Gallego Arias <e+git@x80.org>