test: `dune describe` and virtual libraries
Added a failing test for `dune describe` that involves virtual
libraries.

In the presence of virtual libraries, `dune describe` fails at computing
the right set of libraries, because the code for transitive closure may
detect some overlapping implementations for a given virtual library. In
such a case, the whole set of libraries is dropped.

The detection of overlapping libraries is necessary when the user asks
to build a library or an executable (because `dune` has to pick an
implementation, and this choice should be controlled by the user).

In the case of `dune describe`, however, the goal is to describe the
whole workspace, which may include several implementation for a virtual
library. In this case, if `dune describe` currently fails when several
implementations are found, it cannot fully describe the whole workspace.

The code for transitive closure (when used by `dune describe`) therefore
must not try to detect overlapping implementations.

Signed-off-by: Benoît Montagu <benoit.montagu@inria.fr>